improve PR, add docs and logs
