+++ docs/api/datasets/pyhealth.datasets.TUABDataset.rst
@@ -1,0 +1,19 @@
+﻿pyhealth.datasets.TUABDataset
+===================================
+
+Dataset is available at https://isip.piconepress.com/projects/tuh_eeg/html/downloads.shtml
+
+The TUAB dataset (or Temple University Hospital EEG Abnormal Corpus) is a collection of EEG data acquired at the Temple University Hospital. 
+
+The dataset contains both normal and abnormal EEG readings.
+
+.. autoclass:: pyhealth.datasets.TUABDataset
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+   
+
+   
+   
+
