+++ docs/api/datasets/pyhealth.datasets.TUEVDataset.rst
@@ -1,0 +1,17 @@
+﻿pyhealth.datasets.TUEVDataset
+===================================
+
+Dataset is available at https://isip.piconepress.com/projects/tuh_eeg/html/downloads.shtml
+
+This corpus is a subset of TUEG that contains annotations of EEG segments as one of six classes: (1) spike and sharp wave (SPSW), (2) generalized periodic epileptiform discharges (GPED), (3) periodic lateralized epileptiform discharges (PLED), (4) eye movement (EYEM), (5) artifact (ARTF) and (6) background (BCKG).
+
+.. autoclass:: pyhealth.datasets.TUEVDataset
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+   
+
+   
+   
+
