+++ docs/api/tasks/pyhealth.tasks.temple_university_EEG_tasks.rst
@@ -1,0 +1,8 @@
+﻿pyhealth.tasks.temple_university_EEG_tasks
+=======================================
+
+
+.. automodule:: pyhealth.tasks.temple_university_EEG_tasks
+    :members:
+    :undoc-members:
+    :show-inheritance:
