+++ pyhealth/datasets/tuev.py
@@ -66,8 +66,8 @@ class TUEVDataset(BaseSignalDataset):
         patient_ids = list(set(list(all_files.keys())))
 
         if self.dev:
-            patient_ids = patient_ids[:5]
-            print(patient_ids)
+            patient_ids = patient_ids[:20]
+            # print(patient_ids)
 
         # get patient to record maps
         #    - key: pid:
