+++ pyhealth/tasks/__init__.py
@@ -11,11 +11,8 @@ from .cardiology_detect import (
     cardiology_isCD_fn,
     cardiology_isWA_fn,
 )
-from .EEG_abnormal import (
-    EEG_isAbnormal_fn,
-)
-from .EEG_events import (
-    EEG_events_fn,
+from .temple_university_EEG_tasks import (
+    EEG_isAbnormal_fn, EEG_events_fn,
 )
 from .length_of_stay_prediction import (
     length_of_stay_prediction_eicu_fn,
